requirements.txt
autopep8==1.4.3
entrypoints==0.3
flake8==3.7.7
mccabe==0.6.1
Pillow==6.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
